cryptography/curse_of_the_wren/requirements.txt
enigma-cipher-machine
